# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise Headless CMS API",
    "description": "Enterprise headless CMS backend for multi-site content modeling, editorial workflows, media, localization, delivery APIs, webhooks, API keys, navigation and analytics.",
    "contact": {
      "name": "CMS Platform Team"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tags": {
      "post": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "Create Tag",
        "operationId": "create_tag_api_v1_tags_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "List Tags",
        "operationId": "list_tags_api_v1_tags_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagOut"
                  },
                  "title": "Response List Tags Api V1 Tags Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tags/{tag_id}": {
      "patch": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "Update Tag",
        "operationId": "update_tag_api_v1_tags__tag_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "Delete Tag",
        "operationId": "delete_tag_api_v1_tags__tag_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/menus": {
      "post": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "Create Menu",
        "operationId": "create_menu_api_v1_menus_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "List Menus",
        "operationId": "list_menus_api_v1_menus_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuOut"
                  },
                  "title": "Response List Menus Api V1 Menus Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/menus/{menu_id}/items": {
      "post": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "Add Menu Item",
        "operationId": "add_menu_item_api_v1_menus__menu_id__items_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Menu Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuItemIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "List Menu Items",
        "operationId": "list_menu_items_api_v1_menus__menu_id__items_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Menu Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItemOut"
                  },
                  "title": "Response List Menu Items Api V1 Menus  Menu Id  Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/menus/{menu_id}/items/{item_id}": {
      "patch": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "Update Menu Item",
        "operationId": "update_menu_item_api_v1_menus__menu_id__items__item_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Menu Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuItemIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "Delete Menu Item",
        "operationId": "delete_menu_item_api_v1_menus__menu_id__items__item_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Menu Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/menus/{menu_id}/reorder": {
      "post": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "Reorder Menu",
        "operationId": "reorder_menu_api_v1_menus__menu_id__reorder_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Menu Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItemOut"
                  },
                  "title": "Response Reorder Menu Api V1 Menus  Menu Id  Reorder Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/redirects": {
      "post": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "Create Redirect",
        "operationId": "create_redirect_api_v1_redirects_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedirectIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "List Redirects",
        "operationId": "list_redirects_api_v1_redirects_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RedirectOut"
                  },
                  "title": "Response List Redirects Api V1 Redirects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/redirects/{redirect_id}": {
      "patch": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "Update Redirect",
        "operationId": "update_redirect_api_v1_redirects__redirect_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "redirect_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Redirect Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedirectIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Taxonomy, Navigation & Redirects"
        ],
        "summary": "Delete Redirect",
        "operationId": "delete_redirect_api_v1_redirects__redirect_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "redirect_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Redirect Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MenuIn": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "key": {
            "type": "string",
            "pattern": "^[a-z0-9_-]+$",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "en-US"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "key",
          "name"
        ],
        "title": "MenuIn"
      },
      "MenuItemIn": {
        "properties": {
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "entry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Id"
          },
          "position": {
            "type": "integer",
            "title": "Position",
            "default": 0
          },
          "is_visible": {
            "type": "boolean",
            "title": "Is Visible",
            "default": true
          },
          "open_in_new_tab": {
            "type": "boolean",
            "title": "Open In New Tab",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "label"
        ],
        "title": "MenuItemIn"
      },
...
      "MenuOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          }
        },
        "type": "object",
        "required": [
          "id",
          "site_id",
          "key",
          "name",
          "locale"
        ],
        "title": "MenuOut"
      },
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
...
      "RedirectIn": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "source_path": {
            "type": "string",
            "title": "Source Path"
          },
          "destination_url": {
            "type": "string",
            "title": "Destination Url"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code",
            "default": 301
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "source_path",
          "destination_url"
        ],
        "title": "RedirectIn"
      },
...
      "ReorderIn": {
        "properties": {
          "item_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Item Ids"
          }
        },
        "type": "object",
        "required": [
          "item_ids"
        ],
        "title": "ReorderIn"
      },
...
      "TagIn": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9-]+$",
            "title": "Slug"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "name",
          "slug"
        ],
        "title": "TagIn"
      },
      "TagOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "site_id",
          "name",
          "slug",
          "color"
        ],
        "title": "TagOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}